Backend_Django_restful_api/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
Django==2.2.1
djangorestframework==3.9.4
gensim==3.7.3
lxml==4.3.3
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
requests==2.22.0
scikit-learn==0.20.3
scipy==1.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
urllib3==1.25.2
qcloudapi-sdk-python==2.0.15
trnslation==1.0.5
langdetect == 1.0.7